u-boot: Backport fix related to missing newline in stm32 kconfig
This commit backport a fix from u-boot mainline (3067971aa9ab01cff82d9e4f8f1b776054388c2c)
It fix a u-boot compilation failure that seems to occurs randomly. This issue is cause by
a missing newline in a stm32 kconfig file that prevent the first line of the next kconfig
file to be read properly.

Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@emdalo.com>