fix: Change hss to native hss-payload-generator
Yocto uses the HSS to build U-Boot into a bootable
payload.bin file for PolarFire SoC.

However, the HSS recipe isn't intended to run on
MPFS, and so it should be native to prevent generation
of dbg and dev packages.

Additionally, its actually a HSS payload generator recipe,
not a HSS recipe.