uboot: bump recipe to use U-Boot distro boot feature on the icicle kit
Bump the u-boot-mpfs recipe and update the Icicle Kit to use the distro
boot feature.

Replace the mpfs defined mmcbootpartition env variable with U-boot's
distro_bootpart command, which is defined as part of U-boot's distro
boot functionality. This attempts to standarsize a way for different
distributions to boot in a board-agnostic manner.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>