Add support for the opc-ua industrial edge demo
OPC UA (Open Platform Communications Unified Architecture) is a
machine-to-machine communication protocol, based on a client-server
architecture, where the client initiates the communication and the
server provides the requested data. The communication is structured
around a set of predefined services that define the interaction
between clients and servers.

In this demo, the UAExpert works as OPC UA Client and PolarFire SoC
ICICLE Kit and Video Kit as OPC UA Servers to communicate over OPC
protocol. This demo also explains the ICICLE capability to function
as OPC UA Client and OPC UA Server both.

Signed-off-by: Ahmed Mujtaba Ali <ahmed.ali2@microchip.com>