examples: add python3-flask as a dependency of the iiohttpserver
Python3-flask is required to run the iiohttpserver. Remove the recipe
from the image and add it as a propper dependency instead.

python3-flask-dev will be installed on images that have tools-sdk
IMAGE_FEATURES installed, so it can be safely remove from mpfs-dev-cli
and core-image-minimal-dev list of packages, since it is redundant.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>