recipes: images: tidy-up core-image-minimal-dev image
The core-image-minimal-dev has tools-sdk feature enabled, which means
it will include dev packages for all the recipes in the image. Drop
libgpio-dev from the list, since it is redundant.

Add packagegroup-core-base-utils and remove ncurses-tools, tar, wget,
zip and vim, since they are included in this packagegroup.

Remove vim since there is a vim-tiny package included as part of
packagegroup-core-base-utils.

Drop iw from the recipe list, since this is now selected on a per
machine basis using the MACHINE_FEATURES variable in the machine
configuration file.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>