README: update user instructions following reorg
Detail the new organisation breakdown following the removal of
meta-riscv.
Due to the reorganisation of the Yocto repository, there has been
changes made to the way in which a QSPI image is generated. QSPI images
now have their own machine configurations to keep uniformaty for users
between Yocto and Buildroot configurations. Update the instructions on
how to generate an image for QSPI NAND and QSPI NOR configurations.
Update the link to the Buildroot build system from the now deprecated
Buildroot SDK to the updated Buildroot-External repository.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>