u-boot: icicle-kit-es-auth: load env variables from an env file
For the authenticated boot machine (icicle-kit-es-auth), do not use
the default distro boot provided in the U-boot icicle configuration file.

An authenticated boot image should not have fallback boot mechanisms
for security reasons. Use CONFIG_ENV_SOURCE_FILE to specify a ".env"
file containing the boot commands. This environment file is built-in in
the U-boot binary instead of an external source.
For demonstration purposes, we support one boot target which is mmc.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>