classes: add support for multiple UBI volumes
Add a new bbclass that allows to create an image of type "ubimg"
containing two UBI volumes, one with the U-boot environment and the
other one with a UBI filesystem.

Update the image_type_mtd bbclass to generate a mtdimg image containing
a payload partition and and the ubimage described above in a second
partition. The qspi flash partition layout now looks as shown below:

mtd partition 0: payload
mtd partition 1: UBI with 2 volumes
                  volume 1: contains U-boot environment variable
                  volume 2: contains UBI filesystem

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>