machine: icicle-kit-es-nor: use systemd version 250.5
There is not enough space in the officially supported QSPI nor (Micron
MT25QL256) to fit systemd version 254.4. When using 254.4 most systemd
services failed to start and a "no space left in device" message is
shown while starting systemd.

Use the PREFERRED_VERSION variable in Yocto to apply version 250.5
for icicle-kit-es-nor. This allows to use systemd on this machine.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>