bsp: enable growfs depending on the machine and systemd version used
The growfs systemd services is currently being applied to all systemd
versions and to all target machines. This creates issues when using
an older version of systemd (250.5) in the icicle-kit-es-nor, since
growfs is not available/supported in that systemd version.

In addition to that, this service shouldn't be enabled in the qspi targets
(icicle-kit-es-nand and icicle-kit-es-nor), since the intention is just
to resize rootfs running in MMC/SD with ext4 filesystem.

Rename the systemd bbappend to limit the growfs to systemd version 254.4
and only deploy the fstab with the growfs service to appropiate machines.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>