recipes: base-files: use custom fstab with growfs option on specific machines
Previously, the custom fstab with growfs options was being deployed
to all machines, this was causing the following issue at boot on certain
machines (icicle-kit-es-nand and icicle-kit-es-nor):

[FAILED] Failed to start Grow Root File System.

Only deploy the custom fstab to specific machines for now. Later, we
should evaluate using wks with --fs-options argument.

Deploy the motd to all machines in the BSP as done in previous releases.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>