readme: remove design version info from the readme intro
The PolarFire SoC Yocto BSP readme explicitely mentions the design
version for the Icicle Kit. Having the version in the readme creates
uneccesarry extra maintenance. Remove the design version from the readme,
this information is also found in the release notes published for each
release and in the release announcements.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>