spdx: enable SPDX JSON generation by default
Software Package Data Exchange (SPDX) is an open standard for software bill of
materials (SBOM).

A Software Bill of Materials (SBOM) is a comprehensive inventory that
lists all the packages and libraries included in an application.

SPDX captures versions of components, licenses, copyrights, security
references and other metadata relating to software. SPDX is value in
monitoring the security posture of a distribution. For more information,
see https://spdx.dev or ISO/IEC 5692:2021.

Signed-off-by: Ivan Griffin <ivan.griffin@microchip.com>