TSN Demo - Graphical User Interface
The TSN Demo with GUI involves setting up an Ubuntu machine with
VLAN configurations and deploying the Video Kit by installing the
WIC image and job file. Once the network is configured,
connectivity between the Ubuntu machine and the TSN Demo Video Kit
is verified.

This demo highlights the impact of network congestion and the
effectiveness of TSN in ensuring real-time data transmission
without interruptions.

Signed-off-by: Ahmed Mujtaba Ali <ahmed.ali2@microchip.com>