recipes-bsp, recipes-kernel: set bootargs in u-boot env and use earlycon
Previously, some machines, like the Icicle Kit NAND and NOR, were
setting bootargs in U-Boot, while the rest of the machines were doing
it in the kernel recipe. This change ensures that bootargs are set
consistently in the U-Boot environment across all machines.

The only exception to this ruel for now is the m100pfsevp machine
from the community layer, since the u-boot variables are
handled by Aries embedded.

Also, the earlycon=sbi argument previously passed to the kernel is no
longer applicable, as it requires the SBI 0.1.

earlycon is the recommended bootline argument to use.
The UARTs in the MSS are compatible with the ns16550a standard, and
there is already an earlycon implementation for those. As a result,
the kernel can directly access the UART without going through the SBI.

Replace earlycon=sbi with earlycon in the kernel bootline arguments
for the icicle-kit-es-nand and icicle-kit-es-nor machines.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>